# Swagger Documentation.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "My API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Theme": {
      "get": {
        "tags": [
          "Theme"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Theme"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ThemeAddModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThemeAddModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ThemeAddModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ThemeAddModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Theme/delete": {
      "delete": {
        "tags": [
          "Theme"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ThemeAddModel": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 20,
            "minLength": 2,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}